// src/types/TranslationKey.ts
export enum TranslationKey {
    CONFIRM_MAIL_ADDRESS = 'CONFIRM_MAIL_ADDRESS',
    HELLO_GUEST = 'HELLO_GUEST',
    CONFIRM_MAIL_TEXT = 'CONFIRM_MAIL_TEXT',
    CONFIRM_MAIL_TEXT2 = 'CONFIRM_MAIL_TEXT2',
    SINCERELY = 'SINCERELY',
    DATA_NOT_STORED = 'DATA_NOT_STORED',
    SOMETHING_WRONG = 'SOMETHING_WRONG',
    EMAIL_ALREADY_EXIST = 'EMAIL_ALREADY_EXIST',
    DATA_STORED = 'DATA_STORED',
    SESSION_CREATE_ERROR = 'SESSION_CREATE_ERROR',
    INVALID_TOKEN = 'INVALID_TOKEN',
    TOKEN_EXPIRED = 'TOKEN_EXPIRED',
    SESSION_EXPIRED = 'SESSION_EXPIRED',
    ALREADY_AUTHENTICATED = 'ALREADY_AUTHENTICATED',
    CREDENTIALS_ERROR = 'CREDENTIALS_ERROR',
}